+++ src/utils/columnType.ts
@@ -1,4 +1,13 @@
 import { AST_NODE_TYPES, TSESTree } from '@typescript-eslint/utils';
+import {
+    ArrayTypeNode,
+    LiteralTypeNode,
+    SyntaxKind,
+    TypeChecker,
+    TypeNode,
+    UnionTypeNode,
+    isTypeReferenceNode,
+} from 'typescript';
 import { parseObjectLiteral } from './treeTraversal';
 
 type Column =
@@ -12,7 +21,7 @@ type Column =
 
 type ColumnTypeString = 'string' | 'number' | 'boolean' | 'Date' | 'unknown';
 
-interface ColumnType {
+export interface ColumnType {
     columnType: ColumnTypeString;
     nullable: boolean;
     literal: boolean;
@@ -121,6 +130,7 @@ export function getDefaultColumnTypeForDecorator(column: Column): ColumnParamete
         case 'VersionColumn':
             return { type: 'integer', nullable: false };
         case 'CreateDateColumn':
+        case 'UpdateDateColumn':
             return { type: 'datetime', nullable: false };
         case 'DeleteDateColumn':
             return { type: 'datetime', nullable: true };
@@ -142,6 +152,7 @@ export function convertArgumentToColumnType(
                 if (typeof arg.value === 'string') {
                     return { ...prev, type: arg.value };
                 }
+                return prev;
             default:
                 return prev;
         }
@@ -155,6 +166,83 @@ export function convertArgumentToColumnType(
         literal: false,
         array: parsed.array ?? false,
     };
+}
+
+export function convertTsTypeToColumnType(arg: TypeNode, checker: TypeChecker): ColumnType {
+    switch (arg.kind) {
+        case SyntaxKind.TemplateLiteralType:
+        case SyntaxKind.StringKeyword:
+            return { columnType: 'string', nullable: false, literal: false, array: false };
+
+        case SyntaxKind.NumberKeyword:
+        case SyntaxKind.BigIntKeyword:
+            return { columnType: 'number', nullable: false, literal: false, array: false };
+
+        case SyntaxKind.BooleanKeyword:
+            return { columnType: 'boolean', nullable: false, literal: false, array: false };
+
+        case SyntaxKind.TypeReference:
+            if (isTypeReferenceNode(arg)) {
+                const symbol = checker.getTypeAtLocation(arg.typeName).getSymbol();
+                if (symbol?.getName() === 'Date') {
+                    return { columnType: 'Date', nullable: false, literal: false, array: false };
+                }
+            }
+            break;
+
+        case SyntaxKind.LiteralType: {
+            const literal = arg as LiteralTypeNode;
+            switch (literal.literal.kind) {
+                case SyntaxKind.NullKeyword:
+                    return { columnType: 'unknown', nullable: true, literal: false, array: false };
+
+                case SyntaxKind.StringLiteral:
+                    return { columnType: 'string', nullable: false, literal: true, array: false };
+
+                case SyntaxKind.NumericLiteral:
+                    return { columnType: 'number', nullable: false, literal: true, array: false };
+
+                case SyntaxKind.TrueKeyword:
+                case SyntaxKind.FalseKeyword:
+                    return { columnType: 'boolean', nullable: false, literal: true, array: false };
+
+                default:
+                    break;
+            }
+            break;
+        }
+
+        case SyntaxKind.ArrayType: {
+            const array = arg as ArrayTypeNode;
+            const item = convertTsTypeToColumnType(array.elementType, checker);
+            return { ...item, array: true };
+        }
+
+        case SyntaxKind.UnionType: {
+            const union = arg as UnionTypeNode;
+            return union.types.reduce<ColumnType>(
+                (acc, currentType) => {
+                    const current = convertTsTypeToColumnType(currentType, checker);
+                    return {
+                        columnType:
+                            current.columnType !== 'unknown' ? current.columnType : acc.columnType,
+                        nullable: current.nullable || acc.nullable,
+                        literal: current.literal || acc.literal,
+                        array: current.array || acc.array,
+                    };
+                },
+                {
+                    columnType: 'unknown',
+                    nullable: false,
+                    literal: false,
+                    array: false,
+                },
+            );
+        }
+        default:
+            return { columnType: 'unknown', nullable: false, literal: false, array: false };
+    }
+    return { columnType: 'unknown', nullable: false, literal: false, array: false };
 }
 
 export function convertTypeToColumnType(arg: TSESTree.TypeNode): ColumnType {
@@ -224,10 +312,7 @@ export function convertTypeToColumnType(arg: TSESTree.TypeNode): ColumnType {
 
         case AST_NODE_TYPES.TSArrayType: {
             const item = convertTypeToColumnType(arg.elementType);
-            if (item) {
-                return { ...item, array: true };
-            }
-            return item;
+            return { ...item, array: true };
         }
 
         // TODO: handles these types too